+++ src/app/app.module.ts
@@ -29,6 +29,7 @@ import { BlockGridMixedSelectionDemoComponent } from './blockgrid/blockgrid-mixe
 import { BlockGridMultiSelectionDemoComponent } from './blockgrid/blockgrid-multi-selection.demo';
 import { BlockGridSingleSelectionDemoComponent } from './blockgrid/blockgrid-single-selection.demo';
 import { BreadcrumbDemoComponent } from './breadcrumb/breadcrumb.demo';
+import { BreadcrumbGauntletDemoComponent } from './breadcrumb/breadcrumb-gauntlet.demo';
 import { BubbleDemoComponent } from './bubble/bubble.demo';
 import { BulletDemoComponent } from './bullet/bullet.demo';
 import { BusyIndicatorDemoComponent } from './busyindicator/form.demo';
@@ -255,6 +256,7 @@ import { ButtonsetDemoComponent } from './buttonset/buttonset.demo';
     BlockGridMultiSelectionDemoComponent,
     BlockGridSingleSelectionDemoComponent,
     BreadcrumbDemoComponent,
+    BreadcrumbGauntletDemoComponent,
     BubbleDemoComponent,
     BulletDemoComponent,
     BusyIndicatorDemoComponent,