+++ src/app/breadcrumb/breadcrumb.demo.ts
@@ -1,7 +1,6 @@
 import {
   ChangeDetectionStrategy,
   Component,
-  ElementRef,
   ViewChild
 } from '@angular/core';